feat: Implement legacy protocol processor with unified binary and JSON processing
- Added ProtocolProcessor class for handling protocol operations with fallback mechanisms.
- Introduced ProtocolResult wrapper for operation results.
- Created interfaces for binary and JSON serialization, deserialization, and native calling.
- Implemented processWithFallback method to handle binary and JSON protocols.

feat: Add utility methods for protocol operations

- Introduced ProtocolUtils class with methods for JSON parsing, validation, and conversion.
- Added methods for generating trace IDs, calculating hashes, and sanitizing strings.
- Implemented size limit checks and response creation methods.

feat: Create protocol validation interface and implementation

- Added ProtocolValidator interface for comprehensive data validation.
- Implemented ProtocolValidatorImpl with rules for data type, size, format, and integrity validation.

feat: Develop protocol context management

- Created ProtocolContext class to manage protocol execution state and metadata.
- Implemented builder pattern for flexible context creation.

feat: Implement structured logging for protocol operations

- Added ProtocolLoggerImpl for structured JSON logging of protocol operations.
- Implemented methods for logging operation start, completion, validation results, errors, and metrics.

feat: Enhance protocol validation with XML support

- Extended ProtocolValidatorImpl to include XML validation capabilities.
- Added methods for checking invalid control characters and counting tag occurrences in XML.

+++ src/main/java/com/kneaf/core/performance/EnhancedNativeFloatBuffer.java
@@ -393,7 +393,7 @@ public final class EnhancedNativeFloatBuffer implements AutoCloseable {
             } catch (UnsatisfiedLinkError e) {
                 // Fallback to older API
                 try {
-                    ByteBuffer b = RustPerformance.generateFloatBufferNative(rows, cols);
+                    ByteBuffer b = RustPerformance.generateFloatBufferNative((int)rows, (int)cols);
                     if (b == null) {
                         allocationFailures.incrementAndGet();
                         return null;
